requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==1.1.2
gspread==5.3.2
httplib2==0.20.4
idna==3.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wincertstore==0.2
xmltodict==0.13.0
